+++ examples/restful/app.rb
@@ -2,6 +2,7 @@ require 'sinatra'
 require 'json'
 require 'path'
 require 'finitio'
+require 'rack/robustness'
 
 SCHEMA = Finitio::DEFAULT_SYSTEM.parse (Path.dir/'webspicy/schema.fio').read
 
@@ -20,6 +21,15 @@ disable :show_exceptions
 enable :raise_errors
 
 set :todolist, TODOLIST.dup
+
+use Rack::Robustness do |g|
+  g.no_catch_all
+  g.status 400
+  g.content_type 'application/json'
+  g.body{|ex| { error: ex.message }.to_json }
+  g.on(Finitio::TypeError)
+end
+
 
 post '/reset' do
   settings.todolist = TODOLIST.dup
@@ -53,6 +63,21 @@ get '/todo/:id' do |id|
     {error: "No such todo"}.to_json
   else
     todo.to_json
+  end
+end
+
+patch '/todo/:id' do |id|
+  content_type :json
+  todo = settings.todolist.find{|todo| todo[:id] == Integer(id) }
+  if todo.nil?
+    status 404
+    {error: "No such todo"}.to_json
+  else
+    patch = SCHEMA["TodoPatch"].dress(JSON.load(request.body.read))
+    patched = todo.merge(patch)
+    settings.todolist = settings.todolist.reject{|todo| todo[:id] == Integer(id) }
+    settings.todolist << patched
+    patched.to_json
   end
 end